+++ src/assets_tracking_service/config.py
@@ -131,61 +131,61 @@ class Config:
     class ConfigDumpSafe(TypedDict):
         """Types for `dumps_safe`."""
 
-        version: str
-        db_dsn: str
-        sentry_dsn: str
+        VERSION: str
+        DB_DSN: str
+        SENTRY_DSN: str
         SENTRY_ENABLED: bool
         SENTRY_ENVIRONMENT: str
-        sentry_monitor_slug_ats_run: str
-        enable_provider_geotab: bool
-        enable_provider_aircraft_tracking: bool
-        enabled_providers: list[str]
-        enable_exporter_geojson: bool
-        enable_exporter_arcgis: bool
-        enabled_exporters: list[str]
-        provider_geotab_username: str
-        provider_geotab_password: str
-        provider_geotab_database: str
-        provider_geotab_group_nvs_l06_code_mapping: dict[str, str]
-        provider_aircraft_tracking_username: str
-        provider_aircraft_tracking_password: str
-        provider_aircraft_tracking_api_key: str
-        exporter_geojson_output_path: str
-        exporter_arcgis_username: str
-        exporter_arcgis_password: str
-        exporter_arcgis_item_id: str
+        SENTRY_MONITOR_SLUG_RUN: str
+        ENABLE_PROVIDER_GEOTAB: bool
+        ENABLE_PROVIDER_AIRCRAFT_TRACKING: bool
+        ENABLED_PROVIDERS: list[str]
+        ENALBE_EXPORTER_GEOJSON: bool
+        ENABLE_EXPORTER_ARCGIS: bool
+        ENABLED_EXPORTERS: list[str]
+        PROVIDER_GEOTAB_USERNAME: str
+        PROVIDER_GEOTAB_PASSWORD: str
+        PROVIDER_GEOTAB_DATABASE: str
+        PROVIDER_GEOTAB_GROUP_NVS_LO6_CODE_MAPPING: dict[str, str]
+        PROVIDER_AIRCRAFT_TRACKING_USERNAME: str
+        PROVIDER_AIRCRAFT_TRACKING_PASSWORD: str
+        PROVIDER_AIRCRAFT_TRACKING_API_KEY: str
+        EXPORTER_GEOJSON_OUTPUT_PATH: str
+        EXPORTER_ARCGIS_USERNAME: str
+        EXPORTER_ARCGIS_PASSWORD: str
+        EXPORTER_ARCGIS_ITEM_ID: str
 
     def dumps_safe(self: Self) -> ConfigDumpSafe:
         """Dump config for output to the user with sensitive data redacted."""
         # noinspection PyTestUnpassedFixture
         return {
-            "version": self.version,
-            "db_dsn": self.db_dsn_safe,
-            "sentry_dsn": self.sentry_dsn,
-            "enable_feature_sentry": self.ENABLE_FEATURE_SENTRY,
-            "sentry_environment": self.SENTRY_ENVIRONMENT,
-            "sentry_monitor_slug_ats_run": self.sentry_monitor_slug_ats_run,
-            "enable_provider_geotab": self.ENABLE_PROVIDER_GEOTAB,
-            "enable_provider_aircraft_tracking": self.ENABLE_PROVIDER_AIRCRAFT_TRACKING,
-            "enabled_providers": self.enabled_providers,
-            "enable_exporter_geojson": self.ENABLE_EXPORTER_GEOJSON,
-            "enable_exporter_arcgis": self.ENABLE_EXPORTER_ARCGIS,
-            "enabled_exporters": self.enabled_exporters,
-            "provider_geotab_username": self.PROVIDER_GEOTAB_USERNAME,
-            "provider_geotab_password": self.provider_geotab_password_safe,
-            "provider_geotab_database": self.PROVIDER_GEOTAB_DATABASE,
-            "provider_geotab_group_nvs_l06_code_mapping": self.provider_geotab_group_nvs_l06_code_mapping,
-            "provider_aircraft_tracking_username": self.PROVIDER_AIRCRAFT_TRACKING_USERNAME,
-            "provider_aircraft_tracking_password": self.provider_aircraft_tracking_password_safe,
-            "provider_aircraft_tracking_api_key": self.provider_aircraft_tracking_api_key_safe,
-            "exporter_geojson_output_path": str(self.EXPORTER_GEOJSON_OUTPUT_PATH.resolve()),
-            "exporter_arcgis_username": self.EXPORTER_ARCGIS_USERNAME,
-            "exporter_arcgis_password": self.exporter_arcgis_password_safe,
-            "exporter_arcgis_item_id": self.EXPORTER_ARCGIS_ITEM_ID,
+            "VERSION": self.VERSION,
+            "DB_DSN": self.DB_DSN_SAFE,
+            "SENTRY_DSN": self.SENTRY_DSN,
+            "ENABLE_FEATURE_SENTRY": self.ENABLE_FEATURE_SENTRY,
+            "SENTRY_ENVIRONMENT": self.SENTRY_ENVIRONMENT,
+            "SENTRY_MONITOR_SLUG_RUN": self.SENTRY_MONITOR_SLUG_RUN,
+            "ENABLE_PROVIDER_GEOTAB": self.ENABLE_PROVIDER_GEOTAB,
+            "ENABLE_PROVIDER_AIRCRAFT_TRACKING": self.ENABLE_PROVIDER_AIRCRAFT_TRACKING,
+            "ENABLED_PROVIDERS": self.ENABLED_PROVIDERS,
+            "ENABLE_EXPORTER_GEOJSON": self.ENABLE_EXPORTER_GEOJSON,
+            "ENABLE_EXPORTER_ARCGIS": self.ENABLE_EXPORTER_ARCGIS,
+            "ENABLED_EXPORTERS": self.ENABLED_EXPORTERS,
+            "PROVIDER_GEOTAB_USERNAME": self.PROVIDER_GEOTAB_USERNAME,
+            "PROVIDER_GEOTAB_PASSWORD": self.PROVIDER_GEOTAB_PASSWORD_SAFE,
+            "PROVIDER_GEOTAB_DATABASE": self.PROVIDER_GEOTAB_DATABASE,
+            "PROVIDER_GEOTAB_GROUP_NVS_LO6_CODE_MAPPING": self.PROVIDER_GEOTAB_GROUP_NVS_LO6_CODE_MAPPING,
+            "PROVIDER_AIRCRAFT_TRACKING_USERNAME": self.PROVIDER_AIRCRAFT_TRACKING_USERNAME,
+            "PROVIDER_AIRCRAFT_TRACKING_PASSWORD": self.PROVIDER_AIRCRAFT_TRACKING_PASSWORD_SAFE,
+            "PROVIDER_AIRCRAFT_TRACKING_API_KEY": self.PROVIDER_AIRCRAFT_TRACKING_API_KEY_SAFE,
+            "EXPORTER_GEOJSON_OUTPUT_PATH": str(self.EXPORTER_GEOJSON_OUTPUT_PATH.resolve()),
+            "EXPORTER_ARCGIS_USERNAME": self.EXPORTER_ARCGIS_USERNAME,
+            "EXPORTER_ARCGIS_PASSWORD": self.EXPORTER_ARCGIS_PASSWORD_SAFE,
+            "EXPORTER_ARCGIS_ITEM_ID": self.EXPORTER_ARCGIS_ITEM_ID,
         }
 
     @property
-    def version(self: Self) -> str:
+    def VERSION(self: Self) -> str:
         """
         Application version.
 
@@ -212,14 +212,14 @@ class Config:
         return dsn
 
     @property
-    def db_dsn_safe(self: Self) -> str:
+    def DB_DSN_SAFE(self: Self) -> str:
         """DB_DSN with password redacted."""
         dsn_parsed = dsnparse.parse(self.DB_DSN, EditableDsn)
         dsn_parsed.secret = self._safe_value
         return dsn_parsed.geturl()
 
     @property
-    def sentry_dsn(self) -> str:
+    def SENTRY_DSN(self) -> str:
         """Connection string for Sentry monitoring."""
         return "https://57698b6483c7ac43b7c9c905cdb79943@o39753.ingest.us.sentry.io/4507581411229696"
 
@@ -236,7 +236,7 @@ class Config:
             return self.env.str("SENTRY_ENVIRONMENT", "development")
 
     @property
-    def sentry_monitor_slug_ats_run(self) -> str:
+    def SENTRY_MONITOR_SLUG_RUN(self) -> str:
         """Slug for the Sentry monitor used to track fetch-export runs of the application."""
         return "ats-run"
 
@@ -253,7 +253,7 @@ class Config:
             return self.env.bool("ENABLE_PROVIDER_AIRCRAFT_TRACKING", True)
 
     @property
-    def enabled_providers(self: Self) -> list[str]:
+    def ENABLED_PROVIDERS(self: Self) -> list[str]:
         """List of enabled providers."""
         providers = []
 
@@ -278,7 +278,7 @@ class Config:
             return self.env.bool("ENABLE_EXPORTER_ARCGIS", True)
 
     @property
-    def enabled_exporters(self: Self) -> list[str]:
+    def ENABLED_EXPORTERS(self: Self) -> list[str]:
         """List of enabled exporters."""
         exporters = []
 
@@ -303,7 +303,7 @@ class Config:
             return self.env.str("PASSWORD")
 
     @property
-    def provider_geotab_password_safe(self: Self) -> str:
+    def PROVIDER_GEOTAB_PASSWORD_SAFE(self: Self) -> str:
         """PROVIDER_GEOTAB_PASSWORD with value redacted."""
         return self._safe_value
 
@@ -314,7 +314,7 @@ class Config:
             return self.env.str("DATABASE")
 
     @property
-    def provider_geotab_group_nvs_l06_code_mapping(self: Self) -> dict[str, str]:
+    def PROVIDER_GEOTAB_GROUP_NVS_LO6_CODE_MAPPING(self: Self) -> dict[str, str]:
         """
         Mapping of Geotab group names to NVS L06 codes.
 
@@ -341,7 +341,7 @@ class Config:
             return self.env.str("PASSWORD")
 
     @property
-    def provider_aircraft_tracking_password_safe(self: Self) -> str:
+    def PROVIDER_AIRCRAFT_TRACKING_PASSWORD_SAFE(self: Self) -> str:
         """PROVIDER_AIRCRAFT_TRACKING_PASSWORD with value redacted."""
         return self._safe_value
 
@@ -352,7 +352,7 @@ class Config:
             return self.env.str("API_KEY")
 
     @property
-    def provider_aircraft_tracking_api_key_safe(self: Self) -> str:
+    def PROVIDER_AIRCRAFT_TRACKING_API_KEY_SAFE(self: Self) -> str:
         """PROVIDER_AIRCRAFT_TRACKING_API_KEY with value redacted."""
         return self._safe_value
 
@@ -375,7 +375,7 @@ class Config:
             return self.env.str("PASSWORD")
 
     @property
-    def exporter_arcgis_password_safe(self: Self) -> str:
+    def EXPORTER_ARCGIS_PASSWORD_SAFE(self: Self) -> str:
         """EXPORTER_ARCGIS_PASSWORD with value redacted."""
         return self._safe_value