+++ tests/assets_tracking_service_tests/cli/test_config.py
@@ -18,7 +18,7 @@ class TestCliConfig:
 
         assert result.exit_code == 0
         # 'tes\nt' workaround from https://github.com/pallets/click/issues/1997
-        assert config.db_dsn_safe in result.output.replace("tes\nt", "test")
+        assert config.DB_DSN_SAFE in result.output.replace("tes\nt", "test")
 
     def test_cli_config_check(self: Self, fx_config: Config, fx_cli: CliRunner) -> None:
         """App configuration validates."""