+++ CHANGELOG.md
@@ -45,6 +45,8 @@ Wherever possible, a reference to an issue in the project issue tracker should b
   [#100](https://gitlab.data.bas.ac.uk/MAGIC/assets-tracking-service/-/issues/100)
 * Documenting updated access constraints (public access to latest positions)
   [#46](https://gitlab.data.bas.ac.uk/MAGIC/assets-tracking-service/-/issues/46)
+* Refactoring Sentry related config options into main Config class
+  [#112](https://gitlab.data.bas.ac.uk/MAGIC/assets-tracking-service/-/issues/112)
 
 ## [0.3.5] - 2024-12-09